Dynamic determine the root dir
