+++ src/build.rs
@@ -1,14 +1,9 @@
-use std::{
-    path::{Path, PathBuf},
-    sync::mpsc,
-    time::Duration,
-};
+use std::{path::Path, sync::mpsc, time::Duration};
 
-use crate::{data, entity::Zine, error::ZineError, ZineEngine};
-use anyhow::{anyhow, Context, Result};
+use crate::{data, ZineEngine};
+use anyhow::{Context, Result};
 use notify_debouncer_mini::{new_debouncer, notify::RecursiveMode};
 use tokio::sync::broadcast::Sender;
-use walkdir::WalkDir;
 
 pub async fn watch_build<P: AsRef<Path>>(
     source: P,
@@ -17,7 +12,7 @@ pub async fn watch_build<P: AsRef<Path>>(
     sender: Option<Sender<()>>,
 ) -> Result<()> {
     // Use zine.toml to find root path
-    let (source, zine) = locate_root_zine_folder(std::fs::canonicalize(source)?)?
+    let (source, zine) = crate::locate_root_zine_folder(std::fs::canonicalize(source)?)?
         .with_context(|| "Failed to find the root zine.toml file".to_string())?;
 
     // Also make the dest folder joined in root path?
@@ -87,39 +82,4 @@ fn build(engine: &mut ZineEngine, reload: bool) -> Result<()> {
     engine.build(reload)?;
     println!("Build cost: {}ms", instant.elapsed().as_millis());
     Ok(())
-}
-
-/// Find the root zine file in current dir and try to parse it
-fn parse_root_zine_file<P: AsRef<Path>>(path: P) -> Result<Option<Zine>> {
-    // Find the name in current dir
-    if WalkDir::new(&path).max_depth(1).into_iter().any(|entry| {
-        let entry = entry.as_ref().unwrap();
-        entry.file_name() == crate::ZINE_FILE
-    }) {
-        // Try to parse the root zine.toml as Zine instance
-        return Ok(Some(Zine::parse_from_toml(path)?));
-    }
-
-    Ok(None)
-}
-
-// Locate folder contains the root `zine.toml`, and return path info and Zine instance.
-fn locate_root_zine_folder(path: PathBuf) -> Result<Option<(PathBuf, Zine)>> {
-    match parse_root_zine_file(&path) {
-        Ok(Some(zine)) => return Ok(Some((path, zine))),
-        Err(err) => match err.downcast::<ZineError>() {
-            // Found a root zine.toml, but it has invalid format
-            Ok(inner_err @ ZineError::InvalidRootTomlFile(_)) => return Err(anyhow!(inner_err)),
-            // Found a zine.toml, but it isn't a root zine.toml
-            Ok(ZineError::NotRootTomlFile) => {}
-            // No zine.toml file found
-            _ => {}
-        },
-        _ => {}
-    }
-
-    match path.parent() {
-        Some(parent_path) => locate_root_zine_folder(parent_path.to_path_buf()),
-        None => Ok(None),
-    }
 }
